Extend VerilatedVar to hold ForceableInfo pointer
This commit makes changes to the verilated `__Syms` file
in order to convey additional information about the relation of a
forceable signal to its force control signals without needing to rely on
string concatenation within `verilated_vpi.cpp`. This implementation
builds upon PR #6704 and simplifies the implementation within
`verilated_vpi.cpp`.

This is accomplished by adding a pointer to a `ForceableInfo` class object to
`VerilatedVar`. It contains pointers to the `VerilatedVar`s for the
`__VforceEn` and `__VforceVal` signals, owns a pointer to the
`__VforceRd` signal's `VerilatedVar`, and contains information about the
assignment (continuous or not) of the variable.

It obtains this information through an adapted `varInsert` function,
`forceableVarInsert`, emitted from `V3EmitCSyms.cpp`. There, if a
forceable signal is found, first the control signals are inserted into
the `m_varsp` map and their pointers are given to the the
`forceableVarInsert` function. Within that function, the `ForceableInfo`
object is constructed and ownership given to the base signal's
`VerilatedVar`.

The pointer to `__VforceRd` can be an owning pointer because the
`__VforceRd` signal cannot be set as `public_flat_rw` or
`public_flat_rd`, and so will never be added to the `VerilatedScope`.
Hence, the only instance of it exists in the `VerilatedVar` for its
forceable base signal. The pointers to the `__VforceEn` and
`__VforceVal` signals are already added to the `VerilatedScope` because
they are inherently public, so the base signal's `VerilatedVar`'s
`ForceableInfo` must contain non-owning pointers to them.

This required some changes to the `VerilatedVar`: Its `m_namep` is now a
`std::string` because the name of the `__VforceRd` signal is constructed
at runtime, rather than passed as a `char*` literal as with the other
signals. It has received an owning `unique_ptr` to a `ForceableInfo`
object, which also necessitates re-introducing its move constructor to
allow it to be moved into the `m_varsp` map.