Fix stale pointers in VerilatedImpData::m_hierMap
`VerilatedImp::s()` is a `VerilatedImpData` singleton that contains an
`m_hierMap` whose keys are pointers to `VerilatedScope` objects. Because
it is a singleton, it is not automatically destroyed together with the
`VerilatedModel`. The result of this is that, while the `VerilatedSyms`'
destructor cleans up the `std::vector` _values_ of this map, its _keys_
remain as stale pointers. The consequence of this is that a future
re-construction of the `VerilatedModel` with new scopes adds these _in
addition to_ the existing keys already in the hierarchy map. This can
lead to issues when iterating over the map, or when checking whether a
`VerilatedScope` is present in the map or not.

This bug originally showed itself in the `t_vpi_public_depth_off` test
after adding more member variables to `VerilatedVar`, and only on a
specific GitHub test runner, and only on Ubuntu 22.04. I suspect the
reason it only showed up there is because it is a special case where not
all `VerilatedScope`s are supposed to be present in the map, so upon
reconstruction of the `VerilatedModel`, a `VerilatedScope` that is _not_
supposed to be in the map was placed in the memory location of a
`VerilatedScope` that _was_ supposed to be in the map formerly.

This commit fixes this behavior.