Improve trace initialization code size
Pass const char* to trace pushPrefix. This might seem a bit silly, but
this saves 10% on the total code size in RTLMeter as it eliminates
string constructor/destructor calls in the trace initializations. We
always call this with either string literals or a C string. Helpful
when you are in the -mcmodel=large domain.